Separate Symbol builtin
