+++ boa/src/lib.rs
@@ -54,6 +54,7 @@ pub mod profiler;
 pub mod property;
 pub mod realm;
 // syntax module has a lot of acronyms
+pub mod symbol;
 #[allow(clippy::upper_case_acronyms)]
 pub mod syntax;
 pub mod value;
